Show sorted locales in language and region picker (#51)
The language and region picker was not sorting locales. It just used the order the locales were listed in on the output of the command `localedef --list-archive`. For example this orders en_GB before en_HK, but to the user these are presented as "English, United Kingdom" and "English, Hong Kong". Clearly United Kingdom should not be sorted before Hong Kong.

This patch makes it so the locales as presented to the user are sorted. Screenshot of before and after the patch here: https://imgur.com/a/OjyQN. Before on the left, after (with correct sorting) on the right.

Also added a rstrip() to iteration over locales as otherwise it would add an empty locale for the empty final line of the output of the command `localedef --list-archive` captured in locales. You can see that on the screenshot as the final locale on the left.