Add support for LC_TIME configuration
LC_TIME controls the date format but also the date language. If you were
French and lived in England, your LANGUAGE would be fr but your region would
be en.

If you set LC_TIME to en (like your region, which is what Ubuntu does) your
time and date would not only adopt the English format (consistent with where
you live), it would also show in English, which isn't right since the rest
of your environment is set to show in French.

This is why LC_TIME and LANGUAGE were bound together and set to the same
value.

This commit makes LC_TIME configurable on its own. In certain situations
where the language is the same but the regional settings matter (in Switzerland
for instance), this needs to be configured differently.