Use /etc/locale.conf if it exists
It looks like Debian is starting to support /etc/locale.conf which
is default in other distributions.

This file doesn't exist in Mint 21 and prior versions.
In Mint 22 /etc/default/locale is a symlink towards /etc/locale.conf
so we should write to /etc/locale.conf if it is present.